# swagger.yaml
openapi: 3.0.0
info:
  title: Simple Auth API
  contact: {}
  version: "1.0"
servers:
  - url: https://simple-auth-api.onrender.com/api/v1/auth
    variables: {}
paths:
...
  /delete-account:
    delete:
      tags:
        - User
      summary: deleteAccount
      operationId: deleteAccount
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...